Refactor package-specific methods into their own file, accessed via adapter + use new interface
